Saves item statutes on POST /itemStatuses
* Retrieves items statuses on GET /itemStatuses

+++ resources/migrations/20150821110406-create-item-statuses.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS item_statuses
-(id             bigint,
+(id             serial,
  name           varchar(255),
  mark_as_closed boolean,
  is_default     boolean);
